+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v1.5.0](https://github.com/in2workspace/in2-issuer-api/releases/tag/v1.5.0)
+### Added
+- Compatibility with LEARCredentialMachine to issue LEARCredentialEmployee.
+
 ## [v1.3.0](https://github.com/in2workspace/in2-issuer-api/releases/tag/v1.3.0)
 ### Changed
 - The issuer now issues only LearCredentialEmployee v2.